refactor: remove getUserToken duplication
