+++ src/api/cart/cart.ts
@@ -10,37 +10,6 @@ import type {
     RemoveDiscountCodePayload,
 } from './cart.types';
 import { getCustomerToken } from '../auth/getToken';
-import { getToken } from '../token';
-import { AUTH_TOKEN_KEY } from '../../utilities/constants/constants';
-
-export const getUserToken = () => {
-    let accessToken: string = '';
-
-    const storedToken = getToken(AUTH_TOKEN_KEY);
-    if (storedToken) {
-        accessToken = storedToken;
-    }
-
-    return accessToken;
-};
-
-export const getUserCartByCustomerId = async (
-    customerId: string
-): Promise<CartResponse> => {
-    const token = getUserToken();
-
-    const apiUrl = process.env.CTP_API_URL;
-    const projectKey = process.env.CTP_PROJECT_KEY;
-
-    const productUrl = `${apiUrl}/${projectKey}/carts/customer-id=${customerId}`;
-    const product = await axios.get<CartResponse>(productUrl, {
-        headers: {
-            Authorization: `Bearer ${token}`,
-            'Content-Type': 'application/json',
-        },
-    });
-    return product.data;
-};
 
 export const getCart = async (loginStatus: boolean): Promise<CartResponse> => {
     const token = await getCustomerToken(loginStatus);
