+++ src/api/catalog/getCategories.ts
@@ -1,4 +1,4 @@
-import { getUserToken } from '../product/getProduct';
+import { getUserToken } from '../auth/getToken';
 import { requestGetCategories } from './requestGetCategories';
 
 export const getCategories = async (loginStatus: boolean) => {
